+++ CHANGELOG.md
@@ -8,8 +8,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Added
-- Added my projects page - [#505](https://github.com/DigitalExcellence/dex-frontend/issues/505)
 
+- Added my projects page - [#505](https://github.com/DigitalExcellence/dex-frontend/issues/505)
 - Added project images. - [#464](https://github.com/DigitalExcellence/dex-frontend/issues/464)
 - Added images to highlights - [#463](https://github.com/DigitalExcellence/dex-frontend/issues/463)
 - Updated selection color to improve branding - [#539](https://github.com/DigitalExcellence/dex-frontend/issues/539)
@@ -22,6 +22,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Support multiple call-to-actions. [#454](https://github.com/DigitalExcellence/dex-frontend/issues/454)
 - Refactored content on the homepage - [#502](https://github.com/DigitalExcellence/dex-frontend/issues/502)
 - Changed location of the buttons on the project details page - [#518](https://github.com/DigitalExcellence/dex-frontend/issues/518)
+- Gifs are now supported for project images and highlight images - [#546](https://github.com/DigitalExcellence/dex-frontend/issues/546)
 
 ### Deprecated